backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
redis==5.0.1
